+++ src/mobi/ioio/plotter/shapes/Delay.java
@@ -4,7 +4,7 @@ import mobi.ioio.plotter.CurvePlotter.Curve;
 
 public class Delay implements Curve {
 	private final float pos_[];
-	private final float time_;
+	private final double time_;
 	
 	
 	public Delay(float pos[], float time) {
@@ -13,12 +13,12 @@ public class Delay implements Curve {
 	}
 
 	@Override
-	public float totalTime() {
+	public double totalTime() {
 		return time_;
 	}
 
 	@Override
-	public void getPosTime(float time, float[] xy) {
+	public void getPosTime(double time, float[] xy) {
 		xy[0] = pos_[0];
 		xy[1] = pos_[1];
 	}
